First draft of working bot

+++ lib/adts/include/common.h
@@ -3,6 +3,6 @@
 #ifndef ADTS_COMMON_H
 #define ADTS_COMMON_H
 
-typedef int (*comparator) (void *, void *);
+typedef long (*comparator) (void *, void *);
 
 #endif
